mesh/requirements.txt
absl-py
graph-nets
tensorflow<2
numpy
dm-sonnet<2
tensorflow_probability<0.9
sklearn
dm-tree
matplotlib
gast==0.2.2
meshio
networkx
paraview.simple